+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `Formal\AccessLayer\Driver::sqlite`
 
+### Fixed
+
+- Support for aliased table names when using `Formal\AccessLayer\Query\Delete`
+
 ## 3.0.0 - 2024-07-14
 
 ### Added